Move some tests out of specs project
