+++ test/Microsoft.Extensions.DependencyInjection.Tests/ServiceProviderContainerTests.cs
@@ -8,6 +8,7 @@ using Microsoft.Extensions.DependencyInjection.Specification.Fakes;
 using Microsoft.Extensions.DependencyInjection.Tests.Fakes;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Xunit;
+using Microsoft.Extensions.DependencyInjection.Ordered;
 
 namespace Microsoft.Extensions.DependencyInjection.Tests
 {
@@ -100,6 +101,36 @@ namespace Microsoft.Extensions.DependencyInjection.Tests
             Assert.Equal(
                 $"Cannot instantiate implementation type '{implementationType}' for service type '{serviceType}'.",
                 exception.Message);
+        }
+
+        [Fact]
+        public void NonEnumerableServiceCannotBeIEnumerableResolved()
+        {
+            // Arrange
+            var collection = new ServiceCollection();
+            collection.AddTransient(typeof(IFakeService), typeof(FakeService));
+            var provider = CreateServiceProvider(collection);
+
+            // Act
+            var services = provider.GetService<IEnumerable<IFakeService>>();
+
+            // Assert
+            Assert.Null(services);
+        }
+
+        [Fact]
+        public void IOrderedDoesNotResolveAsIEnumerable()
+        {
+            // Arrange
+            var collection = new ServiceCollection();
+            collection.AddOrdered<IFakeService>().AddTransient<FakeService>();
+            var provider = CreateServiceProvider(collection);
+
+            // Act
+            var ordered = provider.GetService<IEnumerable<IFakeService>>();
+
+            // Assert
+            Assert.Null(ordered);
         }
 
         private abstract class AbstractFakeOpenGenericService<T> : IFakeOpenGenericService<T>
